requirements.txt
Flask-Admin==1.6.1
Flask-Testing==0.8.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.5.0
bleach==6.2.0
click==8.1.8
jsonschema==4.23.0
setproctitle==1.3.4
markupsafe==3.0.2
psycopg2-binary==2.9.5
python-dateutil==2.9.0.post0
ujson==5.10.0
redis==5.2.1
yattag==1.16.1
xmltodict==0.14.2
brainzutils@git+https://github.com/metabrainz/brainzutils-python.git@v2.9.1
spotipy==2.25.1
datasethoster@git+https://github.com/metabrainz/data-set-hoster.git@830ecb2b2120acbd5deed2dab4587784c7be04b6
troi@git+https://github.com/metabrainz/troi-recommendation-playground.git@v2025.01.29.0
PyYAML==6.0.2
eventlet==0.35.2
# eventlet fails to patch dnspython >= 2.3 https://github.com/eventlet/eventlet/issues/781
dnspython==2.6.1
uWSGI==2.0.28
sentry-sdk[flask]==2.19.2
pydantic==1.10.19
pycountry==24.6.1
Flask==3.1.0
Jinja2==3.1.6
Werkzeug==3.1.3
Flask-DebugToolbar==0.16.0
msgpack==0.5.6
requests==2.32.3
SQLAlchemy==2.0.36
mbdata@git+https://github.com/acoustid/mbdata.git@v29.0.0
typesense==0.21.0
unidecode>=1.3.4
Levenshtein==0.26.1
numpy==2.2.1
pandas==2.2.3
pyarrow==18.1.0
more-itertools==10.5.0
kombu==5.4.2
itsdangerous==2.2.0
urllib3==1.26.19
orjson==3.9.15
pyjwt[crypto]==2.7.0
packaging~=23.1
feedgen==1.0.0
WTForms<=3.1.2
requests-oauthlib==2.0.0
